Move initDb to consent resource test
